+++ tests/testthat/test-add-longnostic.R
@@ -1,7 +1,7 @@
 context("test-add-longnostic")
 
 df_add_l_diff_1 <- features(wages_ts, ln_wages, diff)
-df_add_l_n_obs <- add_l_n_obs(wages_ts)
+df_add_n_key_obs <- add_n_key_obs(wages_ts)
 df_add_l_slope <- add_l_slope(wages_ts, ln_wages ~ xp)
 df_add_l_slope_multi <- add_l_slope(wages_ts, ln_wage ~ xp + ged)
 
@@ -9,7 +9,7 @@ updated_dim <- c(nrow(wages), ncol(wages) + 1)
 
 test_that("longnostics returns the right dimensions", {
   expect_equal(dim(df_add_l_diff_1), updated_dim)
-  expect_equal(dim(df_add_l_n_obs), updated_dim)
+  expect_equal(dim(df_add_n_key_obs), updated_dim)
   expect_equal(dim(df_add_l_slope), c(nrow(wages), ncol(wages) + 2))
   expect_equal(dim(df_add_l_slope_multi), c(nrow(wages), ncol(wages) + 3))
 })
@@ -23,7 +23,7 @@ add_new_names <- function(data, x){
 test_that("longnostic returns the right names", {
   expect_equal(names(df_add_l_diff_1), add_new_names(wages,
                                                      "l_diff_1"))
-  expect_equal(names(df_add_l_n_obs), c(names(wages_ts), "n_obs"))
+  expect_equal(names(df_add_n_key_obs), c(names(wages_ts), "n_obs"))
   expect_equal(names(df_add_l_slope), 
                add_new_wage_names(wages_ts,
                                   c("l_intercept",
@@ -37,7 +37,7 @@ test_that("longnostic returns the right names", {
 
 test_that("longnostic returns a tbl_df", {
   expect_is(df_add_l_diff_1, class = c("tbl"))
-  expect_is(df_add_l_n_obs, class = c("tbl"))
+  expect_is(df_add_n_key_obs, class = c("tbl"))
   expect_is(df_add_l_slope, class = c("tbl"))
   expect_is(df_add_l_slope_multi, class = c("tbl"))
 })
@@ -46,7 +46,7 @@ classes <- function(x) purrr::map_chr(x, class)
 
 test_that("longnostic returns correct classes", {
   expect_equal(classes(df_add_l_diff_1)[["l_diff_1"]], "numeric")
-  expect_equal(classes(df_add_l_n_obs)[["n_obs"]], "integer")
+  expect_equal(classes(df_add_n_key_obs)[["n_obs"]], "integer")
   expect_equal(classes(df_add_l_slope)[["l_intercept"]], "numeric")
   expect_equal(classes(df_add_l_slope)[["l_slope_xp"]], "numeric")
   expect_equal(classes(df_add_l_slope_multi)[["l_intercept"]], "numeric")