+++ NEWS.md
@@ -5,6 +5,8 @@
 * renamed `add_k_groups()` to `stratify_keys()`
 * removed many of the `l_<summary>` functions in favour of the `features` approach.
 * rename `l_summarise_fivenum` to `l_summarise`, and have an option to pass a list of functions.
+* rename `l_n_obs()` to `n_key_obs()`
+* rename `l_slope()` to `key_slope()`
 
 # brolgar 0.0.1.9000